[ENH] refactor removing date entity and allow to use bids fitler to only convert certain files from source to raw

+++ src/convertSourceToRaw.m
@@ -1,4 +1,4 @@
-function convertSourceToRaw(cfg)
+function convertSourceToRaw(varargin)
     %
     % Function attempts to convert a source dataset created with CPP_BIDS into a valid
     % BIDS data set.
@@ -6,10 +6,13 @@ function convertSourceToRaw(cfg)
     %
     % USAGE::
     %
-    %   convertSourceToRaw(cfg)
+    %   convertSourceToRaw(cfg, 'filter', filter)
     %
     % :param cfg: cfg structure is needed only for providing the path in ``cfg.dir.output``.
     % :type cfg: structure
+    %
+    % :param filter: bids.query filter to only convert a subset of files.
+    % :type filter: structure
     %
     % :output:
     %          - :creates: a dummy README and CHANGE file
@@ -19,39 +22,80 @@ function convertSourceToRaw(cfg)
     %
     % (C) Copyright 2020 CPP_BIDS developers
 
+    args = inputParser;
+
+    default_filter = struct([]);
+
+    args.addRequired('cfg', @isstruct);
+    args.addParameter('filter', default_filter, @isstruct);
+
+    args.parse(varargin{:});
+
+    cfg = args.Results.cfg;
+    filter = args.Results.filter;
+
     sourceDir = fullfile(cfg.dir.output, 'source');
     rawDir = fullfile(cfg.dir.output, 'raw');
+
+    % back up description to not overwrite
+    % TODO bids malab should be smart enought to not do that
+    isFile = @(x) exist(x, 'file');
+    if isFile(fullfile(rawDir, 'dataset_description.json'))
+        copyfile(fullfile(rawDir, 'dataset_description.json'), ...
+                 fullfile(rawDir, 'dataset_description_bu.json'));
+    end
+
+    % trick use bids matlab copy dataset function
+    bids.copy_to_derivative(sourceDir, ...
+                            'pipeline_name', '.', ...
+                            'out_path', rawDir, ...
+                            'filter', filter, ...
+                            'unzip', false, ...
+                            'force', true, ...
+                            'skip_dep', false, ...
+                            'use_schema', false, ...
+                            'verbose', true);
+
+    if isFile(fullfile(rawDir, 'dataset_description_bu.json'))
+        copyfile(fullfile(rawDir, 'dataset_description_bu.json'), ...
+                 fullfile(rawDir, 'dataset_description.json'));
+        delete(fullfile(rawDir, 'dataset_description_bu.json'));
+    else
+        % clean up description
+        description = bids.util.jsondecode(fullfile(rawDir, 'dataset_description.json'));
+        description.BIDSVersion = '1.7.0';
+        description.Name = 'FIXME';
+        description.DatasetType = 'raw';
+        description = rmfield(description, 'GeneratedBy');
+        description = rmfield(description, 'SourceDatasets');
+        bids.util.jsonencode(fullfile(rawDir, 'dataset_description.json'), description);
+    end
+
+    removeDateEntity(rawDir, 'filter', filter);
+
+    gunzipTimeSeries(rawDir);
 
     % add dummy README and CHANGE file
     templateFolder = fullfile(fileparts(mfilename('fullpath')), '..', 'templates');
 
-    copyfile(fullfile(templateFolder, 'README'), ...
-             sourceDir);
-    copyfile(fullfile(templateFolder, 'CHANGES'), ...
-             sourceDir);
-    copyfile(fullfile(templateFolder, '.bidsignore'), ...
-             sourceDir);
-
-    copyfile(sourceDir, rawDir);
-
-    BIDS = bids.layout(rawDir,  'use_schema', false);
-
-    data = bids.query(BIDS, 'data');
-    metadata = bids.query(BIDS, 'metadata');
-
-    for i = 1:size(data, 1)
-        bf = bids.File(data{i});
-        if isfield(bf.entities, 'date')
-            % TODO probably JSON renaming should be passed to bids-matlab
-            sourceJson = fullfile(fileparts(bf.path), bf.json_filename);
-            bf.entities.date = '';
-            bf.rename('dry_run', false, 'force', true);
-            bids.util.jsonencode(fullfile(fileparts(bf.path), bf.json_filename), metadata{i});
-            delete(sourceJson);
-        end
+    if ~isFile(fullfile(rawDir, 'README'))
+        copyfile(fullfile(templateFolder, 'README'), ...
+                 rawDir);
+    end
+    if ~isFile(fullfile(rawDir, 'CHANGES'))
+        copyfile(fullfile(templateFolder, 'CHANGES'), ...
+                 rawDir);
+    end
+    if ~isFile(fullfile(rawDir, '.bidsignore'))
+        copyfile(fullfile(templateFolder, '.bidsignore'), ...
+                 rawDir);
     end
 
-    BIDS = bids.layout(rawDir,  'use_schema', false);
+end
+
+function gunzipTimeSeries(pathToDataSet)
+
+    BIDS = bids.layout(pathToDataSet,  'use_schema', false);
     data = bids.query(BIDS, 'data', 'suffix', {'stim', 'physio' }, 'ext', '.tsv');
 
     for i = 1:size(data, 1)
@@ -60,5 +104,4 @@ function convertSourceToRaw(cfg)
             delete(data{i});
         end
     end
-
 end
